+++ src/v1/Archetypes/UtilityTypes/EquivalentPart.ts
@@ -32,19 +32,17 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-export * from "./AttributeFilterMap";
-export * from "./AttributeKeys";
-export * from "./AttributeTransformerMap";
-export * from "./Definitely";
-export * from "./EquivalentKeys";
-export * from "./EquivalentOptionalKeys";
-export * from "./EquivalentOptionalPart";
-export * from "./EquivalentPart";
-export * from "./IdenticallyNamedKeys";
-export * from "./IdenticallyNamedPart";
-export * from "./IfEquals";
-export * from "./Maybe";
-export * from "./OptionalKeys";
-export * from "./OptionalPart";
-export * from "./WritableKeys";
-export * from "./WritablePart";
+
+import { EquivalentKeys } from "./EquivalentKeys";
+
+/**
+ * `EquivalentPart` is a mapped type. Use it to create an interface that
+ * only contains the fields of both `A` and `B` that are equivalent types.
+ *
+ * See {@link EquivalentKeys} for details about how we decide which fields
+ * in `A` and `B` are equivalent.
+ *
+ * @category UtilityTypes
+ */
+export type EquivalentPart<A extends object, B extends object> =
+    Pick<A, EquivalentKeys<A,B>>;